+++ src/index.html
@@ -13,6 +13,6 @@
   </head>
   <body>
     <!-- ❗️❗️❗️ Add partials like so for every page -->
-    <!-- <include src="./partials/example.html"></include> -->
+    <include src="./partials/main-section.html"></include>
   </body>
 </html>
