+++ src/templates/filmCards.hbs
@@ -3,8 +3,8 @@
     <img src='{{poster_path}}' alt='{{title}}' class='film__item' />
     <p class='film__title'>{{title}}</p>
     <p class='film__info'>
-      {{genre_ids}}
+      {{genres}}
       <span class='film__year'>|{{release_date}}</span>
     </p>
   </a>
-</li>;
+</li>
